+++ RELEASE_NOTES.md
@@ -1,3 +1,1 @@
- * ```nmap``` is always required
- * added ```%SIGALGO%``` to the possible output formatting
- * added ```--default-format``` to display the default formatting output
+Fixed the parsing of the signature algorithm